backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
supabase==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
